content/canvas/test/test_2d.state.saverestore.stackdepth.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.state.saverestore.stackdepth</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var limit = 512;
for (var i = 1; i < limit; ++i)
{
    ctx.save();
    ctx.lineWidth = i;
}
for (var i = limit-1; i > 0; --i)
{
    ok(ctx.lineWidth == i, "ctx.lineWidth == i");
    ctx.restore();
}

if (!_deferred) SimpleTest.finish();
});
</script>
